refactor(util): move dedup report script to existing folder
Ref. metriport/metriport-internal#2179
