+++ packages/utils/src/fhir-deduplication/report/report-dedup.ts
@@ -6,8 +6,8 @@ import { getEnvVarOrFail } from "@metriport/core/util/env-var";
 import { out } from "@metriport/core/util/log";
 import fs from "fs";
 import { groupBy } from "lodash";
-import { ellapsedTimeAsStr } from "../shared/duration";
-import { initRunsFolder } from "../shared/folder";
+import { ellapsedTimeAsStr } from "../../shared/duration";
+import { initRunsFolder } from "../../shared/folder";
 import { processAllergyIntolerance } from "./allergy-intolerance";
 import { processCondition } from "./condition";
 import { processCoverage } from "./coverage";
